+++ src/module.ts
@@ -2,8 +2,9 @@ import { Module } from '@nuxt/types'
 
 interface Options {
   tokenProperty: String
-  refreshTokenProperty: String
   tokenType: String
+  refreshTokenProperty?: String
+  rememberProperty?: String
   loginPath?: String
   guestPath?: String
   prefix?: String
@@ -13,19 +14,25 @@ interface Options {
   refreshUrl?: String
   registerUrl?: String
   userUrl?: String
+  userDataProperty?: String
   editUrl?: String
 }
 
 const myModule: Module<Options> = function (moduleOptions) {
   const { resolve, join } = require('path')
+
   // const { readdirSync } = require('fs')
   const options = {
     ...moduleOptions,
     ...this.options.authCustom,
   }
-
+  if (!options.tokenProperty)
+    throw new Error(`[nuxt-auth-jy] Option "tokenProperty" doesn't set`)
+  if (!options.tokenType)
+    throw new Error(`[nuxt-auth-jy] Option "tokenType" doesn't set`)
   if (!options.loginPath) options.loginPath = '/'
   if (!options.guestPath) options.guestPath = '/'
+  if (!options.userDataProperty) options.userDataProperty = 'data'
   if (!options.prefix) options.prefix = 'authCustom_'
   if (!options.i18n) options.i18n = false
   if (!options.expires) options.expires = 3600